src/requirements.txt
Django==1.9
EasyProcess==0.2.1
PyVirtualDisplay==0.1.5
argparse==1.2.1
decorator==4.0.6
djangorestframework==3.3.2
ipython==4.0.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
selenium==2.48.0
simplegeneric==0.8.1
traitlets==4.0.0
wsgiref==0.1.2
psycopg2
gunicorn
